feat(web): basic play-game API stub

+++ internal/game/board_random.go
@@ -26,12 +26,12 @@ func CreateRandomlyPlannedBoard() (*Board, error) {
 	return board, nil
 }
 
-func placePieces(board *Board, unplaced map[string]*pieces.Piece, startRow uint8) (*Board, error) {
+func placePieces(board *Board, unplaced map[string]*pieces.Piece, startRow int) (*Board, error) {
 	numColumns := int(board.GetSize().Columns)
 	inx := 0
 	for _, p := range unplaced {
-		row := uint8(inx / numColumns)
-		column := uint8(inx - int(row)*numColumns)
+		row := int(inx / numColumns)
+		column := int(inx - int(row)*numColumns)
 		row += startRow
 		err := board.PlacePiece(p, Position{R: row, C: column})
 
@@ -39,8 +39,8 @@ func placePieces(board *Board, unplaced map[string]*pieces.Piece, startRow uint8
 			log.Error().
 				Err(err).
 				Int("number", inx).
-				Uint8("row", row).
-				Uint8("number", column).
+				Int("row", row).
+				Int("number", column).
 				Str("piece", p.GetRank().String()).
 				Str("position", fmt.Sprintf("(%d, %d)", row, column)).
 				Msg("failed to place piece")
